requirements.txt

python-telegram-bot==20.3
mediapipe
opencv-python-headless
torch
torchvision
numpy
